Fix VS 2015 windows compilation: conversion from 'const unsigned int' to 'int' requires a narrowing conversion

+++ src/openpose/hand/handExtractor.cpp
@@ -220,7 +220,7 @@ namespace op
                 if (!mHeatMapTypes.empty())
                 {
                     mHeatMaps[0].reset({numberPeople, (int)HAND_NUMBER_PARTS, mNetOutputSize.y, mNetOutputSize.x});
-                    mHeatMaps[1].reset({numberPeople, HAND_NUMBER_PARTS, mNetOutputSize.y, mNetOutputSize.x});
+                    mHeatMaps[1].reset({numberPeople, (int)HAND_NUMBER_PARTS, mNetOutputSize.y, mNetOutputSize.x});
                 }
 
                 // // Debugging
